pyproject.toml
[project]
name = "kontur-api"
version = "0.1.0"
description = "Desktop automation tool for Kontur.Marking workflows"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "customtkinter==5.2.2",
    "openpyxl==3.1.5",
    "pandas==2.3.2",
    "python-dotenv==1.1.1",
    "pywebview>=4.4",
    "requests==2.32.5",
    "selenium==4.35.0",
    "trustme>=1.2",
    "pywin32==311; platform_system == 'Windows'",
]

[dependency-groups]
dev = [
    "flake8==7.3.0",
    "mypy==1.18.2",
    "types-requests==2.32.4.20250913",
]

[tool.uv]
package = false
default-groups = []
